requirements.txt
certifi==2018.10.15
chardet==3.0.4
defusedxml==0.5.0
dj-database-url==0.5.0
Django==2.1.3
django-allauth==0.38.0
django-disqus==0.5
django-tagging==0.4.6
gunicorn==19.9.0
idna==2.7
mock==2.0.0
oauthlib==2.1.0
pbr==5.1.1
Pillow==5.3.0
psycopg2-binary==2.7.6.1
python3-openid==3.1.0
pytz==2018.7
requests==2.20.1
requests-oauthlib==1.0.0
six==1.11.0
urllib3==1.24.1
whitenoise==4.1.2
